add jest testing
